fix: move my-poor-room component to page folder
