--- src/my-poor-room/components/challenge-achievement/index.ts
@@ -1,4 +1,0 @@
-import { ChallengeAchievement } from './ChallengeAchievement';
-import { ProgressBar } from './Progressbar';
-
-export { ChallengeAchievement, ProgressBar };
