--- src/my-poor-room/components/challenge-categories/index.ts
@@ -1,4 +1,0 @@
-import { ChallengeCategories } from './ChallengeCategories';
-import { ChallengeCategory } from './ChallengeCategory';
-
-export { ChallengeCategories, ChallengeCategory };
