--- src/my-poor-room/components/chip/index.ts
@@ -1,3 +1,0 @@
-import { DateChip } from './DateChip';
-
-export { DateChip };
