+++ src/pages/my-poor-room/ChallengeCategory.tsx
