+++ src/pages/my-poor-room/index.tsx
@@ -2,10 +2,11 @@ import { useRouter } from 'next/router';
 import { ReactElement } from 'react';
 
 import { ChallengeRoomFeedList, MyRoomFeedList } from '@/features/feed';
-import { ChallengeAchievement } from '@/my-poor-room/components/challenge-achievement';
-import { ChallengeCategories } from '@/my-poor-room/components/challenge-categories';
 import { Spacing } from '@/shared/components';
 import BottomNavLayout from '@/shared/components/layout/BottomNavLayout';
+
+import { ChallengeAchievement } from './ChallengeAchievement';
+import { ChallengeCategories } from './ChallengeCategories';
 
 export default function MyPoorRoom() {
   const router = useRouter();
